pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tribe-setup"
version = "0.1.0"
description = "Minimal runner wrapper around Meta's Tribe v2 model."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "moviepy>=2.2.1",
  "nilearn>=0.11,<1",
  "numpy>=2.2,<3",
  "pandas>=2.2,<3",
  "PyYAML>=6.0,<7",
  "tribev2 @ git+https://github.com/facebookresearch/tribev2.git",
]

[project.scripts]
tribe-cli = "tribe_cli.main:main"
neurokit = "tribe_cli.main:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
]

[tool.pytest.ini_options]
pythonpath = [".", "packages/sdk/src", "packages/cli/src"]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = [".", "packages/sdk/src", "packages/cli/src"]
